Prototype for model settings object

Basic R6 object for Settings object, a List that stores settings used related to model training.

Public fields

name

The default name of this settings as character.

modelid

A character of the model id this belongs to.

data

A list of contained settings.

Methods

Public methods

Settings$new()

Settings$print()

Settings$show()

Settings$length()

Settings$duration()

Settings$summary()

Settings$get()

Settings$set()

Settings$clone()

Method new()

Initializes the object and creates an empty list

Usage

Settings$new()

Returns

NULL

Method print()

Print the names and properties of all Biodiversity datasets contained within

Usage

Settings$print()

Returns

A message on screen

Method show()

Shows the name and the settings

Usage

Settings$show()

Returns

A character of the name and settings.

Method length()

Number of options

Usage

Settings$length()

Returns

A numeric with the number of options.

Method duration()

Computation duration convenience function

Usage

Settings$duration()

Returns

The amount of time passed for model fitting if found.

Method summary()

Summary call of the contained parameters

Usage

Settings$summary()

Returns

A list with the parameters in this object.

Method get()

Get a specific setting

Usage

Settings$get(what)

Arguments

what

A character with the respective setting.

Returns

The setting if found in the object.

Method set()

Set new settings

Usage

Settings$set(what, x, copy = FALSE)

Arguments

what

A character with the name for the new settings.

x

The new setting to be stored. Can be any object.

copy

logical on whether a new settings object is to be created.

Returns

The setting if found in the object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Settings$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.