Add biodiversity polygon dataset to a distribution object (presence-only)

This function can be used to add a sf polygon dataset to an existing distribution object. Presence-only polygon data is treated differential than point data in subsequent engines and models.

Usage

add_biodiversity_polpo(
  x,
  polpo,
  name = NULL,
  field_occurrence = "Observed",
  formula = NULL,
  family = "poisson",
  simulate = FALSE,
  simulate_points = 100,
  simulate_weights = NULL,
  simulate_strategy = "random",
  separate_intercept = TRUE,
  ...
)

# S4 method for BiodiversityDistribution,sf
add_biodiversity_polpo(x, polpo)

Arguments

x

distribution() (i.e. BiodiversityDistribution) object.

polpo

A sf or Spatial) polygon object of presence-only occurrences.

name

The name of the biodiversity dataset used as internal identifier

field_occurrence

A numeric or character location of biodiversity point records.

formula

A character or formula object to be passed. Default is to use all covariates (if specified)

family

A character stating the family to be used (Default: Poisson)

simulate

Simulate poipa points within its boundaries. Result are passed to add_biodiversity_poipa (Default: FALSE)

simulate_points

A numeric number of points to be created by simulation

simulate_weights

A Raster layer describing an eventual preference for simulation (Default: NULL)

simulate_strategy

A character stating the strategy for sampling. Can be set to either 'random' or 'regular', the latter requiring a raster supplied in the simulate_weights parameter.

separate_intercept

A boolean value stating whether a separate intercept is to be added in shared likelihood models for engines engine_inla, engine_inlabru and engine_stan.

...

Other parameters passed down

Details

The default approach for polygon data is tosample presence-only points across the region of the polygons. This function thus adds as a wrapper to add_biodiversity_poipo as presence-only points are created by the model. Although engine_inla and engine_inlabru in the ibis.iSDM package are able to include (small) polygon data directly in the modelling, the link between covariates and polygonal data is established through a nearest-neighbour matching and thus is relatively equivalent to a point sampling.

The 'simulate' options allow specify parameters with regards to the simulation steps.

For an integration of range data as predictor or offset, see add_predictor_range and add_offset_range instead.

Notes

References

See also

Examples

if (FALSE) {
 TBD
}