Adds a threshold constraint to a scenario object

This option adds a threshold() constraint to a scenario projection, thus effectively applying the threshold as mask to each projection step made during the scenario projection.

Applying this constraint thus means that the "suitability" projection is clipped to the threshold. This method requires the threshold() set for a scenario object.

It could be in theory possible to re calculate the threshold for each time step based on supplied parameters or even observation records. So far this option has not been necessary to implement.

Usage

add_constraint_threshold(mod, updatevalue = NA, ...)

# S4 method for class 'BiodiversityScenario'
add_constraint_threshold(mod, updatevalue = NA, ...)

Arguments

mod

A BiodiversityScenario object with specified predictors.

updatevalue

A numeric indicating to what the masked out values (those outside) the threshold should become (Default: NA).

...

passed on parameters. See also the specific methods for adding constraints.

Note

Threshold values are taken from the original fitted model.

See also

Examples

if (FALSE) { # \dontrun{
# Add scenario constraint
scenario(fit) |> threshold() |>
add_constraint_threshold()
} # }