Add GLOBIOM-DownScaleR derived predictors to a Biodiversity distribution object

This is a customized function to format and add downscaled land-use shares from the Global Biosphere Management Model (GLOBIOM) to a distribution or BiodiversityScenario in ibis.iSDM. GLOBIOM is a partial-equilibrium model developed at IIASA and represents land-use sectors with a rich set of environmental and socio-economic parameters, where for instance the agricultural and forestry sector are estimated through dedicated process-based models. GLOBIOM outputs are spatial explicit and usually at a half-degree resolution globally. For finer grain analyses GLOBIOM outputs can be produced in a downscaled format with a customized statistical downscaling module.

The purpose of this script is to format the GLOBIOM outputs of DownScale for the use in the ibis.iSDM package.

Usage

add_predictors_globiom(
  x,
  fname,
  names = NULL,
  transform = "none",
  derivates = "none",
  derivate_knots = 4,
  int_variables = NULL,
  bgmask = TRUE,
  harmonize_na = FALSE,
  priors = NULL,
  ...
)

# S4 method for class 'BiodiversityDistribution,character'
add_predictors_globiom(
  x,
  fname,
  names = NULL,
  transform = "none",
  derivates = "none",
  derivate_knots = 4,
  int_variables = NULL,
  bgmask = TRUE,
  harmonize_na = FALSE,
  priors = NULL,
  ...
)

# S4 method for class 'BiodiversityScenario,character'
add_predictors_globiom(
  x,
  fname,
  names = NULL,
  transform = "none",
  derivates = "none",
  derivate_knots = 4,
  int_variables = NULL,
  bgmask = TRUE,
  harmonize_na = FALSE,
  priors = NULL,
  ...
)

Arguments

x

A BiodiversityDistribution or BiodiversityScenario object.

fname

A character pointing to a netCDF with the GLOBIOM data.

names

A vector of character names describing the environmental stack in case they should be renamed (Default: NULL).

transform

A vector stating whether predictors should be preprocessed in any way (Options: 'none','pca', 'scale', 'norm')

derivates

A Boolean check whether derivate features should be considered (Options: 'none', 'thresh', 'hinge', 'quad') )

derivate_knots

A single numeric or vector giving the number of knots for derivate creation if relevant (Default: 4).

int_variables

A vector with length greater or equal than 2 specifying the covariates (Default: NULL).

bgmask

Check whether the environmental data should be masked with the background layer (Default: TRUE)

harmonize_na

A logical value indicating of whether NA values should be harmonized among predictors (Default: FALSE)

priors

A PriorList object. Default is set to NULL which uses default prior assumptions.

...

Other parameters passed down

Details

See add_predictors() for additional parameters and customizations. For more (manual) control the function for formatting the GLOBIOM data can also be called directly via formatGLOBIOM().

See also

Examples

if (FALSE) { # \dontrun{
 obj <- distribution(background) |>
        add_predictors_globiom(fname = "", transform = 'none')
 obj
} # }