Create a new bdproto object

Construct a new object with bdproto. This object system is inspired from the ggproto system used in the ggplot2 and the custom proto system used in the prioritizr package.

Usage

bdproto(`_class` = NULL, `_inherit` = NULL, ...)

Source

https://ggplot2.tidyverse.org/reference/ggproto.html#examples

Arguments

_class

Class name to assign to the object. This is stored as the class attribute of the object. This is optional: if NULL (the default), no class name will be added to the object.

_inherit

bdproto object to inherit from. If NULL, don't inherit from any object.

...

A list of members to add to the new bdproto object.

Value

A bdproto object.

Note

All bdproto objects have several functions that allow to access data and summarize information within them. Run names(object) to show the available functions.

See also

Examples

Adder <- bdproto("Adder",
  x = 0,
  add = function(self, n) {
    self$x <- self$x + n
    self$x
  }
)
Adder$add(10)
#> [1] 10
Adder$add(10)
#> [1] 20
Abacus <- bdproto("Abacus", Adder,
  subtract = function(self, n) {
    self$x <- self$x - n
    self$x
  }
)
Abacus$add(10)
#> [1] 30
Abacus$subtract(10)
#> [1] 20