Engine for Bayesian regularized regression models

Efficient MCMC algorithm for linear regression models that makes use of 'spike-and-slab' priors for some modest regularization on the amount of posterior probability for a subset of the coefficients.

Usage

engine_breg(
  x,
  iter = 10000,
  nthread = getOption("ibis.nthread"),
  type = "response",
  ...
)

Arguments

x

distribution() (i.e. BiodiversityDistribution) object.

iter

numeric on the number of MCMC iterations to run (Default: 10000).

type

The mode used for creating posterior predictions. Either making "link" or "response" (Default: "response").

...

Other none specified parameters passed on to the model.

nthreads

numeric on the number of CPU-threads to use for data augmentation.

Details

This engine provides efficient Bayesian predictions through the Boom R-package. However note that not all link and models functions are supported and certain functionalities such as offsets are generally not available. This engines allows the estimation of linear and non-linear effects via the "only_linear" option specified in train.

References

Nguyen, K., Le, T., Nguyen, V., Nguyen, T., & Phung, D. (2016, November). Multiple kernel learning with data augmentation. In Asian Conference on Machine Learning (pp. 49-64). PMLR.

Steven L. Scott (2021). BoomSpikeSlab: MCMC for Spike and Slab Regression. R package version 1.2.4. https://CRAN.R-project.org/package=BoomSpikeSlab

See also