Use INLA as engine

References https://becarioprecario.bitbucket.io/inla-gitbook/ch-spatial.html#ref-Simpsonetal:2016

engine_inla(
  x,
  optional_mesh = NULL,
  optional_projstk = NULL,
  max.edge = c(1, 5),
  offset = c(1, 1),
  cutoff = 1,
  proj_stepsize = NULL,
  timeout = NULL,
  barrier = FALSE,
  ...
)

Arguments

x
distribution() (i.e. BiodiversityDistribution) object.

optional_mesh
A directly supplied INLA mesh (Default: NULL)

optional_projstk
A directly supplied projection stack. Useful if projection stack is identical for multiple species (Default: NULL)

max.edge
The largest allowed triangle edge length, must be in the same scale units as the coordinates

offset
interpreted as a numeric factor relative to the approximate data diameter;

cutoff
The minimum allowed distance between points on the mesh

proj_stepsize
The stepsize in coordinate units between cells of the projection grid (Default: NULL)

timeout
Specify a timeout for INLA models in sec. Afterwards it passed.

barrier
Should a barrier model be added to the model?

...
Other variables

nonconvex.bdry
Create a non-convex boundary hulls instead (Default: FALSE) TBD

nonconvex.convex
Non-convex minimal extension radius for convex curvature TBD

nonconvex.concave
Non-convex minimal extension radius for concave curvature TBD

nonconvex.res
Computation resolution for nonconvex.hulls TBD