Use inlabru as engine

Model components are specified with general inputs and mapping methods to the latent variables, and the predictors are specified via general R expressions, with separate expressions for each observation likelihood model in multi-likelihood models. The inlabru engine - similar as the engine_inla function acts a wrapper for INLA::inla albeit inlabru has a number of convenience functions implemented that make particular predictions more straight forward (e.g. via posterior sampling).

Usage

engine_inlabru(
  x,
  optional_mesh = NULL,
  optional_projstk = NULL,
  max.edge = c(1, 5),
  offset = c(1, 1),
  cutoff = 1,
  proj_stepsize = NULL,
  timeout = NULL,
  strategy = "auto",
  int.strategy = "auto",
  ...
)

Source

https://inlabru-org.github.io/inlabru/articles/

Arguments

x

distribution() (i.e. BiodiversityDistribution) object.

optional_mesh

A directly supplied INLA mesh (Default: NULL)

optional_projstk

A directly supplied projection stack. Useful if projection stack is identical for multiple species (Default: NULL)

max.edge

The largest allowed triangle edge length, must be in the same scale units as the coordinates

offset

interpreted as a numeric factor relative to the approximate data diameter;

cutoff

The minimum allowed distance between points on the mesh

proj_stepsize

The stepsize in coordinate units between cells of the projection grid (Default: NULL)

timeout

Specify a timeout for INLA models in sec. Afterwards it passed.

strategy

Which aproximation to use for the joint posterior. Options are "auto","adaptative", "gaussian", "simplified.laplace" & "laplace".

int.strategy

Integration strategy. Options are "auto","grid", "eb" & "ccd".

...

Other variables

References

Bachl, F. E., Lindgren, F., Borchers, D. L., & Illian, J. B. (2019). inlabru: an R package for Bayesian spatial modelling from ecological survey data. Methods in Ecology and Evolution, 10(6), 760-766.

https://becarioprecario.bitbucket.io/inla-gitbook/ch-spatial.html#ref-Simpsonetal:2016