ibis.iSDM

Integrated framework of modelling the distribution of species and ecosystems in a suitability framing. This package allows the estimation of integrated species distribution models (iSDM) based on several sources of evidence and provided presence-only and presence-absence datasets. It makes heavy use of point-process models for estimating habitat suitability and allows to include spatial latent effects and priors in the estimation. To do so 'ibis.iSDM' supports a number of engines for Bayesian and more non-parametric machine learning estimation. Further, the 'ibis.iSDM' is specifically customized to support spatial-temporal projections of habitat suitability into the future.

See also

Author

Maintainer: Martin Jung jung@iiasa.ac.at (ORCID) [copyright holder]

Other contributors:

Maximilian H.K. Hesselbarth hesselbarth@iiasa.ac.at (ORCID) [contributor]