Print ibis options

There are a number of hidden options that can be specified for ibis.iSDM. Currently supported are:

'ibis.runparallel' : logical value on whether processing should be run in parallel.

'ibis.nthread' : numeric value on how many cores should be used by default.

'ibis.setupmessages' : logical value indicating whether message during object creation should be shown (Default: NULL).

'ibis.engines' : Returns a vector with all valid engines.

'ibis.use_future' : logical on whether the future package should be used for parallel computing.

Usage

ibis_options()

Value

The output of getOptions for all ibis related variables.

Examples

ibis_options()
#> $ibis.cleannames
#> [1] TRUE
#> 
#> $ibis.corPred
#> [1] 0.7
#> 
#> $ibis.dependencies
#>  [1] "pdp"           "scales"        "biscale"       "modEvA"       
#>  [5] "dplyr"         "geodist"       "geosphere"     "progress"     
#>  [9] "glmnet"        "glmnetUtils"   "xgboost"       "BoomSpikeSlab"
#> [13] "INLA"          "inlabru"       "gnlm"          "cubelyr"      
#> [17] "matrixStats"   "Boruta"        "abess"         "gdalUtilities"
#> [21] "dbarts"        "mboost"        "rstan"         "cmdstanr"     
#> [25] "biscale"       "poems"         "BiocManager"  
#> 
#> $ibis.engines
#>  [1] "GDB-Model"     "BART-Model"    "INLABRU-Model" "BREG-Model"   
#>  [5] "GLMNET-Model"  "GLM-Model"     "SCAMPR-Model"  "INLA-Model"   
#>  [9] "STAN-Model"    "XGBOOST-Model"
#> 
#> $ibis.futurestrategy
#> [1] "sequential"
#> 
#> $ibis.nthread
#> [1] 3
#> 
#> $ibis.priors
#> [1] "INLAPrior"   "BARTPrior"   "GDBPrior"    "GLMNETPrior" "XGBPrior"   
#> [6] "BREGPrior"   "STANPrior"  
#> 
#> $ibis.pseudoabsence
#> Background Settings: 5 parameters
#> 
#> $ibis.runparallel
#> [1] FALSE
#> 
#> $ibis.seed
#> [1] 14143
#> 
#> $ibis.setupmessages
#> [1] TRUE
#>