Create a polynomial transformation from coordinates

This function transforms the coordinates of a supplied file through a polynomial transform. By default it applies weights and a QR decomposition for numerical stability.

polynominal_transform(
  coords,
  degree = 2,
  weights = rep(1/nrow(coords), nrow(coords))
)

Arguments

coords
A data.frame, matrix or sf object with coordinates (2 columns named x-y)

degree
The number of degrees used for polynominal transformation (Default: 2)

weights
Set by default to the inverse of the number of coordinates.

Value

A data.frame with transformed coordinates.

References

Dray S., Plissier R., Couteron P., Fortin M.J., Legendre P., Peres-Neto P.R., Bellier E., Bivand R., Blanchet F.G., De Caceres M., Dufour A.B., Heegaard E., Jombart T., Munoz F., Oksanen J., Thioulouse J., Wagner H.H. (2012). Community ecology in the age of multivariate multiscale spatial analysis. Ecological Monographs 82, 257–275.