Creates a new PriorList object

A PriorList object is essentially a list that contains individual Prior objects. In order to use priors for any of the engines, the respective Prior has to be identified (e.g. INLAPrior) and embedded in a PriorList object. Afterwards these objects can then be added to a distribution object with the add_priors function.

A PriorList object is essentially a list that contains individual Prior objects. In order to use priors for any of the engines, the respective Prior has to be identified (e.g. INLAPrior) and embedded in a PriorList object. Afterwards these objects can then be added to a distribution object with the add_priors function.

Usage

priors(x, ...)

# S4 method for class 'ANY'
priors(x, ...)

priors(x, ...)

# S4 method for class 'ANY'
priors(x, ...)

Arguments

x

A Prior object added to the list.

...

One or multiple additional Prior object added to the list.

Value

A PriorList object.

A PriorList object.

See also

Examples

p1 <- GDBPrior(variable = "Forest", hyper = "positive")
p2 <- GDBPrior(variable = "Urban", hyper = "decreasing")
priors(p1, p2)
#> Set priors: 2

if (FALSE) { # \dontrun{
p1 <- INLAPrior(variable = "Forest",type = "normal", hyper = c(1,1e4))
p2 <- INLAPrior(variable = "Urban",type = "normal", hyper = c(0,1e-2))
priors(p1, p2)
} # }