render_html

Renders DistributionModel to HTML

Usage

render_html(mod, file, title = NULL, author = NULL, notes = "-", ...)

# S4 method for ANY
render_html(mod, file, title = NULL, author = NULL, notes = "-", ...)

Arguments

mod

Any object belonging to DistributionModel

file

Character with path to file.

title

Character with title of document.

author

Character with name of author.

notes

Character with notes added at the beginning of the document.

...

Currently not used

Value

Writes HTML file

Details

Renders a HTML file with several summaries of a trained DistributionModel. The file paths must be an HTML file ending. The functions creates a temporary Rmd file that gets renders as a HTML using the file argument.

Examples

if (FALSE) {
mod <- distribution(background) |>
  add_biodiversity_poipo(species) |>
  add_predictors(predictors) |>
  engine_glmnet() |>
  train()

render_html(mod, file = "Test.html")
}