Remove existing priors from an existing distribution object

This function allows to remove priors from an existing distribution object. In order to remove a set prior, the name of the prior has to be specified.

Usage

rm_priors(x, names = NULL, ...)

# S4 method for BiodiversityDistribution
rm_priors(x, names = NULL, ...)

Arguments

x

distribution (i.e. BiodiversityDistribution) object.

names

A vector or character object for priors to be removed.

...

Other parameters passed down

See also

Examples

if (FALSE) {
 # Add prior
 pp <-  GLMNETPrior("forest")
 x <- distribution(background) |>
  add_priors(pp)
 # Remove again
 x <- x |> rm_priors("forest")
}