Add priors to an existing distribution object

This function simply allows to add priors to an existing distribution object. The supplied priors must be a PriorList object created through calling priors.

Usage

# S4 method for class 'BiodiversityDistribution'
set_priors(x, priors = NULL, ...)

Arguments

x

distribution (i.e. BiodiversityDistribution) object.

priors

A PriorList object containing multiple priors.

...

Other parameters passed down.

Note

Alternatively priors to environmental predictors can also directly added as parameter via add_predictors

See also

Examples

if (FALSE) { # \dontrun{
 pp <-  GLMNETPrior("forest")
 x <- distribution(background) |>
  add_priors(pp)

} # }