Calculate environmental similarity of reference datasets to predictors.

Calculate the environmental similarity of the provided covariates with respect to a reference dataset. Currently supported is Multivariate Environmental Similarity index and the multivariate combination novelty index (NT2) based on the Mahalanobis divergence (see references).

similarity(obj, ref, ref_type = "poipo", method = "mess", ...)

# S4 method for BiodiversityDistribution
similarity(obj)

# S4 method for RasterBrick
similarity(obj)

# S4 method for RasterStack
similarity(obj)

Arguments

obj
A BiodiversityDistribution, DistributionModel or alternatively a Raster object

ref
A BiodiversityDistribution, DistributionModel or alternatively a data.frame with extracted values (corresponding to those given in obj)

ref_type
A character specifying the type of biodiversity to use when obj is a BiodiversityDistribution

method
A specifc method for similarity calculation. Currently supported: 'mess', 'nt'

full
should similarity values be returned for all variables (Default: FALSE)?

plot
Should the result be plotted? Otherwise return the output list (Default: TRUE)

Value

This function returns a list containing:

similarity: a RasterStack giving the environmental similarities for each variable in x (only included when full=TRUE);

mis: a Raster layer giving the minimum similarity value across all variables for each location (i.e. the MESS);

exip: a Raster layer indicating whether any model would interpolate or extrapolate to this location based on environmental surface;

mod: a factor Raster layer indicating which variable was most dissimilar to its reference range (i.e. the MoD map, Elith et al. 2010); and

mos: a factor Raster layer indicating which variable was most similar to its reference range.

Details

similarity implements the MESS algorithm described in Appendix S3 of Elith et al. 2010 as well as the Mahalanobis dissimilarity described in Mesgaran et al. 2014

References

Elith, J., Kearney, M., and Phillips, S. (2010) The art of modelling range-shifting species. Methods in Ecology and Evolution, 1: 330-342. doi:10.1111/j.2041-210X.2010.00036.x

Mesgaran, M.B., Cousens, R.D. and Webber, B.L. (2014) Here be dragons: a tool for quantifying novelty due to covariate range and correlation change when projecting species distribution models. Diversity and Distributions, 20: 1147-1159. doi:10.1111/ddi.12209

Examples

 if (FALSE) {
library(dismo)
library(raster)
ff <- list.files(system.file('ex', package='dismo'), '\\.grd$',
                 full.names=TRUE )
predictors <- stack(grep('biome', ff, value=TRUE, invert=TRUE))
occ <- read.csv(system.file('ex/bradypus.csv', package='dismo'))[, -1]
ref <- extract(predictors, occ)
mess <- similarity(predictors, ref, full=TRUE)
plot
 }
if (FALSE) {
library(rasterVis)
library(RColorBrewer)
levelplot(mess[[2]], col.regions=brewer.pal(8, 'Set1'))
}