Threshold a continuous prediction to a categorical layer

For BiodiversityScenario objects store a threshold attribute in the scenario object

Usage

threshold(
  obj,
  method = "mtp",
  value = NULL,
  poi = NULL,
  return_threshold = FALSE,
  ...
)

# S4 method for ANY
threshold(obj)

# S4 method for RasterBrick
threshold(obj)

# S4 method for RasterLayer
threshold(obj)

# S4 method for BiodiversityScenario
threshold(obj)

Arguments

obj

A trained DistributionModel or alternatively a Raster object

method

A specifc method for thresholding. One of 'fixed', 'mtp', 'percentile'

value

A numeric value for thresholding if method is fixed (Default: NULL)

poi

A sf object containing observational data used for model training

return_threshold

Should threshold value be returned instead (Default: FALSE)

Details

The following options are currently implemented:

'fixed' = applies a single pre-determined threshold

'mtp' = minimum training presence find and sets the lowest predicted suitability for any occurrence point

'percentile' = For a percentile threshold

'TSS' = Determines the optimal TSS (True Skill Statistic). Requires the 'modEvA' package

'kappa' = Determines the optimal kappa value (Kappa). Requires the 'modEvA' package

'F1score' = Determines the optimal F1score (also known as Sorensen similarity). Requires the 'modEvA' package

'F1score' = Determines the optimal sensitivity of presence records. Requires the 'modEvA' package

'Sensitivity' = Determines the optimal sensitivity of presence records. Requires the 'modEvA' package

'Specificity' = Determines the optimal sensitivity of presence records. Requires the 'modEvA' package

References

Lawson, C.R., Hodgson, J.A., Wilson, R.J., Richards, S.A., 2014. Prevalence, thresholds and the performance of presence-absence models. Methods Ecol. Evol. 5, 54–64. https://doi.org/10.1111/2041-210X.12123

Liu, C., White, M., Newell, G., 2013. Selecting thresholds for the prediction of species occurrence with presence-only data. J. Biogeogr. 40, 778–789. https://doi.org/10.1111/jbi.12058

Examples

if (FALSE) {
print('test')
}