Train the model from a given engine

TODO: Add explanation See #details for particularities on how parameters are handled by the individual engines

Note several engines do not support joint likelihood models Therefore the strategy is to either run the model in sequence and informing the model as highlighted in Miller et al. 2019. In this case, when multiple datasets are supplied, the model is run in sequence where the nth' dataset is provided as input to the nth'+1 dataset.

Usage

train(
  x,
  runname,
  rm_corPred = FALSE,
  varsel = "reg",
  inference_only = FALSE,
  only_linear = TRUE,
  bias_variable = NULL,
  bias_value = NULL,
  verbose = FALSE,
  ...
)

# S4 method for BiodiversityDistribution
train(x)

Arguments

x

distribution() (i.e. BiodiversityDistribution) object).

runname

A character name of the trained run

rm_corPred

Remove highly correlated predictors. Default is True

varsel

Perform a variable selection on the set of predictors either prior to building the model or via variable selection / regularization of the model. Available options are none, reg either through AIC / regularization or hyperparameter tuning (Default), and abess as adaptive best subset selection via the abess R-package.

inference_only

Fit model only without spatial projection (Default: FALSE)

only_linear

Fit model only on linear covariate baselearners (DEFAULT: TRUE)

bias_variable

A vector with names of variables to be set to bias_value (Default: NULL)

bias_value

A vector with values to be set to bias_variable (Default: NULL)

...

further arguments passed on.

Value

A distribution prediction object

Details

Train a distribution() model with the specified engine.

TBD: Further details

References

Miller, D.A.W., Pacifici, K., Sanderlin, J.S., Reich, B.J., 2019. The recent past and promising future for data integration methods to estimate species’ distributions. Methods Ecol. Evol. 10, 22–37. https://doi.org/10.1111/2041-210X.13110

See also