Helper function when multiple variables are supplied for a BREG prior

This is a helper function to specify several BREGPrior with the same hyper-parameters, but different variables.

Usage

BREGPriors(variable, hyper = NULL, ip = NULL)

# S4 method for class 'character'
BREGPriors(variable, hyper = NULL, ip = NULL)

Arguments

variable

A character matched against existing predictors.

hyper

A numeric estimate of the mean regression coefficients.

ip

A numeric estimate between 0 and 1 of the inclusion probability of the target variable (Default: NULL).

See also