Project a fitted model to a new environment and covariates

Equivalent to train, this function acts as a wrapper to project the model stored in a BiodiversityScenario object to newly supplied (future) covariates. Supplied predictors are usually spatial-temporal predictors which should be prepared via add_predictors() (e.g. transformations and derivates) in the same way as they have been during the initial modelling with distribution(). Any constrains specified in the scenario object are applied during the projection.

Usage

# S4 method for BiodiversityScenario, character, logical, character, character
project(mod, date_interpolation, stabilize, stabilize_method, layer)

Arguments

mod

A BiodiversityScenario object with set predictors. Note that some constrains such as MigClim can still simulate future change without projections.

date_interpolation

A character on whether dates should be interpolated. Options include "none" (Default), "annual", "monthly", "daily".

stabilize

A boolean value indicating whether the suitability projection should be stabilized (Default: FALSE).

stabilize_method

character stating the stabilization method to be applied. Currently supported is `loess`.

layer

A character specifying the layer to be projected (Default: "mean").

...

passed on parameters.

Value

Saves stars objects of the obtained predictions in mod.

Details

In the background the function x$project() for the respective model object is called, where x is fitted model object. For specifics on the constrains, see the relevant constrain functions, respectively:

add_constrain() for generic wrapper to add any of the available constrains.

add_constrain_dispersal() for specifying dispersal constrain on the temporal projections at each step.

add_constrain_MigClim() Using the MigClim R-package to simulate dispersal in projections.

add_constrain_connectivity() Apply a connectivity constrain at the projection, for instance by adding a barrier that prevents migration.

add_constrain_adaptability() Apply an adaptability constrain to the projection, for instance constraining the speed a species is able to adapt to new conditions.

add_constrain_boundary() To artificially limit the distribution change. Similar as specifying projection limits, but can be used to specifically constrain a projection within a certain area (e.g. a species range or an island).

Many constrains also requires thresholds to be calculated. Adding threshold() to a BiodiversityScenario object enables the computation of thresholds at every step based on the threshold used for the main model (threshold values are taken from there).

Finally this function also allows temporal stabilization across prediction steps via enabling the parameter stabilize and checking the stablize_method argument. Stabilization can for instance be helpful in situations where environmental variables are quite dynamic, but changes in projected suitability are not expected to abruptly increase or decrease. It is thus a way to smoothen out outliers from the projection. Options are so far for instance 'loess' which fits a loess() model per pixel and time step. This is conducted at the very of the processing steps and any thresholds will be recalculated afterwards.

See also