Specify settings for absence point sampling over the model background

This function defines the settings for pseudo-absence sampling of the background. For many engines such points are necessary to model poisson (or binomial) distributed data. Specifically we call absence points for binomial (bernoulli really) distributed responses 'pseudo-absence' and absence data for Poisson responses 'background' points. For more details read Renner et al. (2015).

The function 'add_pseudoabsence' allows to add absence points to any sf object.

Usage

# S4 method for ANY, numeric, numeric, character, numeric, logical, ANY
pseudoabs_settings(background, nrpoints, min_ratio, method, buffer_distance, mcp_inside, bias)

Arguments

background

A RasterLayer or sf object over which background points can be sampled. Default is NULL (Default) and the background is then added when the sampling is first called.

nrpoints

A numeric given the number of background points to be created (Default: 10 000).

min_ratio

A numeric with the minimum ratio of background points relative to presence points. Usually ignored unless the ratio exceeds the nrpoints parameters (Default: 0.25).

method

character denoting how the sampling should be done. See details for options (Default: "random").

buffer_distance

numeric A distance from the observations in which pseudo-absence points are not to be generated. Note that units follow the units of the projection (e.g. m or °).

mcp_inside

A logical value of whether absence points should be sampled outside (Default) or inside the minimum convex polygon provided this method is chosen (parameter method = "mcp").

bias

A RasterLayer with the same extent and projection and background. Absence points will be preferentially sampled in areas with higher (!) bias. (Default: NULL).

...

Any other settings to be added to the pseudoabs settings.

Details

Mandatory and possible parameters are:

background Specifies the extent over which absence points are to be sampled.

nrpoints The number of points to be generated.

method The specific method on how absence points should be generated. Available options are 'random' (Default), 'buffer' to generate only within a buffered distance of existing points, or 'mcp' to only generate them within or outside a minimum convex polygon of the presence points. The parameter mcp_inside supports inside vs outside generation.

buffer_distance The numeric value indicating the distance from presence points where absence points are to be created.

min_ratio The minimum ratio of absence points relative presence points. Ensures a minimum number of absence points. For instance setting this value to 1 generates an equal amount of absence points relative to presence points.

bias An optional bias layer over which points should preferentially be sampled. Points are preferentially generated in areas with larger bias.

References

Renner IW, Elith J, Baddeley A, Fithian W, Hastie T, Phillips SJ, Popovic G, Warton DI. 2015. Point process models for presence-only analysis. Methods in Ecology and Evolution 6:366–379. DOI: 10.1111/2041-210X.12352.

Examples

if (FALSE) {
# It is also possible to match the number of presence-absence points directly.
pseudoabs_settings(nrpoints = 0, min_ratio = 1)
}