Train the model from a given engine

Train a distribution() model with the specified engine.

train(
  x,
  runname,
  rm_corPred = FALSE,
  varsel = FALSE,
  inference_only = FALSE,
  only_linear = TRUE,
  bias_variable = NULL,
  bias_value = NULL,
  verbose = FALSE,
  ...
)

# S4 method for BiodiversityDistribution
train(x)

Arguments

x
distribution() (i.e. BiodiversityDistribution) object).

runname
A character name of the trained run

rm_corPred
Remove highly correlated predictors. Default is True

varsel
Perform a variable selection on the set of predictors

inference_only
Fit model only without spatial projection (Default: FALSE)

only_linear
Fit model only on linear covariate baselearners (DEFAULT: TRUE)

bias_variable
A vector with names of variables to be set to bias_value (Default: NULL)

bias_value
A vector with values to be set to bias_variable (Default: NULL)

...
further arguments passed on.

Value

A distribution prediction object

Details