Train the model from a given engine

This function trains a distribution() model with the specified engine and furthermore has some generic parameters that apply to all engines (regardless of type).

Users are advised to check the help files for individual engine for advice on how the estimation is being done.

Usage

train(
  x,
  runname,
  rm_corPred = FALSE,
  varsel = "reg",
  inference_only = FALSE,
  only_linear = TRUE,
  bias_variable = NULL,
  bias_value = NULL,
  verbose = FALSE,
  ...
)

# S4 method for BiodiversityDistribution
train(x)

Arguments

x

distribution() (i.e. BiodiversityDistribution) object).

runname

A character name of the trained run.

rm_corPred

Remove highly correlated predictors (Default: FALSE). This option removes - based on pairwise comparisons - those covariates that are highly collinear (Pearson's r = >0.7).

varsel

Perform a variable selection on the set of predictors either prior to building the model or via variable selection / regularization of the model. Available options are:

none for no or default priors and no extensive hyperparameter search.

reg Model selection either through DIC or regularization / hyperparameter tuning depending on the engine (Default option).

abess A-priori adaptive best subset selection of covariates via the abess package (see References). Note that this effectively fits a separate generalized linear model to reduce the number of covariates. Can be helpful for engines that don't directly support efficient variable regularization and when N>100.

inference_only

By default the engine is used to create a spatial prediction of the suitability surface, which can take time. If only inferences of the strength of relationship between covariates and observations are required, this parameter can be set to TRUE to ignore any spatial projection (Default: FALSE).

only_linear

Fit model only on linear baselearners and functions. Depending on the engine setting this option to FALSE will result in non-linear relationships between observations and covariates, often increasing processing time (Default: TRUE). How non-linearity is captured depends on the used engine.

bias_variable

A vector with names of variables to be set to bias_value (Default: NULL). This option can for instance be used to 'partial' out certain biases after predictions have been made. See Examples.

bias_value

A vector with values to be set to bias_variable (Default: NULL). Specifying a numeric value here sets bias_variable to the target value.

...

further arguments passed on.

Value

A DistributionModel object.

Details

The resulting object contains both a fit_best object of the estimated model and, if inference_only is FALSE a RasterLayer object named prediction that contains the spatial prediction of the model. These objects can be requested via object$get_data("fit_best").

Note that several engines do not support fully integrated likelihood models (such engine_gdb or engine_xgboost) for example. In this case the estimation strategy for multiple added biodiversity datasets (e.g. more than 1) is to fit recurrent models in sequence, where the nth'} spatial projection is provided as input to the \code{nth'+1 dataset. See also Miller et al. (2019) in the references for more details on this strategy. Of course, if users want more control about this aspect, another option is to fit separate models and make use of the add_offset_range and ensemble functionalities.

References

Miller, D.A.W., Pacifici, K., Sanderlin, J.S., Reich, B.J., 2019. The recent past and promising future for data integration methods to estimate species’ distributions. Methods Ecol. Evol. 10, 22–37. https://doi.org/10.1111/2041-210X.13110

Zhu, J., Wen, C., Zhu, J., Zhang, H., & Wang, X. (2020). A polynomial algorithm for best-subset selection problem. Proceedings of the National Academy of Sciences, 117(52), 33117-33123.

See also

Examples

if (FALSE) {
 # Fit a linear penalized logistic regression model via stan
 x <- distribution(background) %>%
        # Presence-absence data
        add_biodiversity_poipa(surveydata) %>%
        # Add predictors and scale them
        add_predictors(env = predictors, transform = "scale", derivates = "none") %>%
        # Use stan for estimation
        engine_stan(chains = 2, iter = 1000, warmup = 500)
 # Train the model
 mod <- train(x, only_linear = TRUE, varsel = 'reg')
 mod
}