Validation of distribution object

This function conducts a comprehensive model evaluation based on either on the fitted point data or any supplied independent. Currently only supporting point datasets. For validation of integrated models more work is needed.

Usage

# S4 method for ANY, character, sf, character, character
validate(mod, method, point, layer, point_column)

# S4 method for RasterLayer, character, sf, character
validate(mod, method, point, point_column)

Arguments

mod

A fitted BiodiversityDistribution object with set predictors. Alternatively one can also provide directly a RasterLayer, however in this case the point layer also needs to be provided.

method

Should the validation be conducted on continuous metrics or thresholded? See Details.

layer

In case multiple layers exist, which one to use? (Default: 'mean').

point

A sf object with type POINT or MULTIPOINT.

point_column

A character vector with the name of the column containing the independent observations. (Default: 'observed').

...

Other parameters that are passed on. Currently unused.

Value

Return a tidy tibble with validation results.

Details

The validate function does not work for all datasets equally.

Note

If you use the Boyce Index, cite the original Hirzel et al. (2006) paper.

References

Liu, C., White, M., Newell, G., 2013. Selecting thresholds for the prediction of species occurrence with presence-only data. J. Biogeogr. 40, 778–789. https://doi.org/10.1111/jbi.12058

Hirzel, A. H., Le Lay, G., Helfer, V., Randin, C., & Guisan, A. (2006). Evaluating the ability of habitat suitability models to predict species presences. Ecological modelling, 199(2), 142-152.

Examples

if (FALSE) {
 # Assuming that mod is a distribution object and has a thresholded layer
 mod <- threshold(mod, method = "TSS")
 validate(mod, method = "discrete")
 }